Introduced Node class.
Created a class for rappresenting a Node of the list.
Redisegned the header of the main class, the next step will be implementing the new function.
I decide to keep easy the public methods, so the main three are:
push for add an element
get for getting an element
remove for removing an element.